Updated failing tests: syntaxTests/{enums, indexing}.

+++ test/libsolidity/syntaxTests/enums/literal_conversion.sol
@@ -3,7 +3,7 @@ contract C {
     function f() public pure {
         Test a = Test(0);
         Test b = Test(1);
-        Test c = Test(type(uint).max);
+        Test c = Test(type(uint256).max);
         a; b; c;
     }
 }
